GH-14: OpenSearch/Cognito terraform work
Added cognito group, and my user, to terraform.

Cleaned up doc/infrastructure.md which has a list of things manually done
to get everything working.  Now that there's a practical opensearch
terraform module, some of this won't need to be done manually again.
Making progress on it, and that progress is reflected in the doc updates.